Moved all prediction code into predict.py predict.py contains code for generating the reconstructed images from VQVAE, and generation of new images using PixelCNN.
README.md heading added to ensure it is properly in the github directory.

Removed prediction code from train.py
