+++ recognition/s4581401-VQVAE/README.md
@@ -1,0 +1,41 @@
+#Generative VQVAE using PixelCNN on the ADNI Dataset
+
+---
+
+## Overview
+
+---
+
+## Models
+
+---
+
+###VQVAE
+
+###Pixel CNN
+
+##Data pre-processing
+
+---
+
+##Training the model
+
+---
+
+##Results
+
+---
+
+##Improvements
+
+---
+
+##Dependencies
+
+---
+
+##References
+
+---
+
+<p align="center"><img src=""></p>
